Separating navigation stack in diff files for a better understanding
