+++ src/navigation/MainNavigator.js
@@ -1,0 +1,69 @@
+import { createStackNavigator } from "@react-navigation/stack";
+import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
+import { Ionicons } from "@expo/vector-icons";
+
+import ChatListScreen from "../screens/ChatListScreen";
+import ChatSettingsScreen from "../screens/ChatSettingsScreen";
+import SettingsScreen from "../screens/SettingsScreen";
+
+const Stack = createStackNavigator();
+const Tab = createBottomTabNavigator();
+
+const TabNavigator = () => {
+    return (
+        <Tab.Navigator>
+            <Tab.Screen
+                name="ChatList"
+                component={ChatListScreen}
+                options={{
+                    tabBarLabel: "Chats",
+                    tabBarIcon: ({ color, size }) => (
+                        <Ionicons
+                            name="chatbubble-outline"
+                            size={size}
+                            color={color}
+                        />
+                    ),
+                }}
+            />
+            <Tab.Screen
+                name="Settings"
+                component={SettingsScreen}
+                options={{
+                    tabBarLabel: "Settings",
+                    tabBarIcon: ({ color, size }) => (
+                        <Ionicons
+                            name="settings-outline"
+                            size={size}
+                            color={color}
+                        />
+                    ),
+                }}
+            />
+        </Tab.Navigator>
+    );
+};
+
+const MainNavigator = (props) => {
+    return (
+        <Stack.Navigator>
+            <Stack.Screen
+                name="Home"
+                component={TabNavigator}
+                options={{
+                    headerShown: false,
+                }}
+            />
+            <Stack.Screen
+                name="Settings"
+                component={ChatSettingsScreen}
+                options={{
+                    headerTitle: "Settings",
+                    headerBackTitle: "Back",
+                }}
+            />
+        </Stack.Navigator>
+    );
+};
+
+export default MainNavigator;
